refactor: move bench responsiveness types to homeboy-extension-contract
Same data/behavior split as the gate cut: BenchResponsivenessSummary and
BenchFailureMemorySample are pure serde types (all primitive fields) with a
pure impl (responsiveness_lost). Their module's core dependencies
(ExtensionChildResourceSummary, Error, Path, env-var reads) live only in the
functions (read_responsiveness_summary, summarize_responsiveness_pings,
memory_sample, missed_ping_window_ms), which stay in core.

Core re-exports the types (pub use). This clears the last sibling field
dependency of BenchScenario - all of BenchScenario/BenchRunSnapshot's field
types are now in contract crates.
