+++ crates/homeboy-core/src/extension/bench/responsiveness.rs
@@ -4,28 +4,11 @@ use serde::{Deserialize, Serialize};
 
 use crate::engine::resource::ExtensionChildResourceSummary;
 use crate::error::{Error, Result};
+pub use homeboy_extension_contract::bench_responsiveness::{
+    BenchFailureMemorySample, BenchResponsivenessSummary,
+};
 
 const DEFAULT_MISSED_PING_WINDOW_MS: u64 = 10_000;
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(deny_unknown_fields)]
-pub struct BenchResponsivenessSummary {
-    pub missed_ping_count: u64,
-    pub max_ping_gap_ms: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub last_ping_at: Option<String>,
-    pub ping_count: u64,
-    pub missed_ping_window_ms: u64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(deny_unknown_fields)]
-pub struct BenchFailureMemorySample {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sampled_peak_rss_bytes: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sampled_peak_cpu_percent: Option<f64>,
-}
 
 #[derive(Debug, Clone, Deserialize)]
 struct BenchResponsivenessPing {
@@ -33,12 +16,6 @@ struct BenchResponsivenessPing {
     at: Option<String>,
     #[serde(default)]
     t_ms: Option<u64>,
-}
-
-impl BenchResponsivenessSummary {
-    pub fn responsiveness_lost(&self) -> bool {
-        self.missed_ping_count > 0
-    }
 }
 
 pub fn missed_ping_window_ms() -> u64 {
