+++ crates/homeboy-extension-contract/src/lib.rs
@@ -14,6 +14,7 @@ pub mod autofix_config;
 pub mod bench_artifact;
 pub mod bench_diagnostics;
 pub mod bench_gate;
+pub mod bench_responsiveness;
 pub mod bench_result;
 pub mod capability;
 pub use bench_artifact::{BenchArtifact, BenchArtifactViewer, BenchPreviewLifecycleMetadata};
@@ -22,6 +23,7 @@ pub use bench_diagnostics::{
     BenchPhaseSummary,
 };
 pub use bench_gate::{BenchGate, BenchGateOp, BenchGateResult};
+pub use bench_responsiveness::{BenchFailureMemorySample, BenchResponsivenessSummary};
 pub use bench_result::{
     BenchChildCommandFailure, BenchMemory, BenchMetricDirection, BenchMetricPhase,
     BenchMetricPolicy, BenchMetrics, BenchProvenance, BenchProvenanceLink, BenchRunExecution,
